Implements API controllers for admin routes
We've decided to move the admin management functionality to React-Admin.
This will be implemented in a separate React based app that will
connect to this application and use it as an API. React-Admin expects
APIs to be formatted in a certain way for it to show it's full power.
Therefore here we have refactored or where necessary implemented
the following APIs:
- Fact Check
- News
- Users