Introduces a news submission form
The form has been refactored to take into account news fields on the
News model. Furthermore, it was designed being inspired by taking
exemple on other fact checking platforms.
We also added code to drop all every indexes on News because of
the fact that slug in not an item of News anymore.